// migrations/snapshots/1e32dc4f2fa2e2da52ed756e72a4a311a5e153e8a5beb8107081ec4f8d26076d/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Club: {
      readonly clubId: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly branch: CodecTypes['pg/text@1']['output'] | null;
      readonly website: CodecTypes['pg/text@1']['output'] | null;
      readonly facebook: CodecTypes['pg/text@1']['output'] | null;
      readonly twitter: CodecTypes['pg/text@1']['output'] | null;
      readonly instagram: CodecTypes['pg/text@1']['output'] | null;
      readonly linkedin: CodecTypes['pg/text@1']['output'] | null;
      readonly facultyInchargeId: CodecTypes['pg/int4@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly Event: {
      readonly eventId: CodecTypes['pg/int4@1']['output'];
      readonly eventName: CodecTypes['pg/text@1']['output'];
      readonly eventDescription: CodecTypes['pg/text@1']['output'] | null;
      readonly aicteHr: CodecTypes['pg/int4@1']['output'] | null;
      readonly date: CodecTypes['pg/timestamptz-temporal@1']['output'] | null;
      readonly time: CodecTypes['pg/text@1']['output'] | null;
      readonly venue: CodecTypes['pg/text@1']['output'] | null;
      readonly hostId: CodecTypes['pg/int4@1']['output'] | null;
      readonly banner: CodecTypes['pg/text@1']['output'] | null;
      readonly brochure: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly Ticket: {
      readonly ticketId: CodecTypes['pg/int4@1']['output'];
      readonly eventId: CodecTypes['pg/int4@1']['output'];
      readonly userId: CodecTypes['pg/int4@1']['output'] | null;
      readonly tnpUid: CodecTypes['pg/text@1']['output'] | null;
      readonly date: CodecTypes['pg/timestamptz-temporal@1']['output'] | null;
      readonly time: CodecTypes['pg/text@1']['output'] | null;
      readonly venue: CodecTypes['pg/text@1']['output'] | null;
      readonly eventName: CodecTypes['pg/text@1']['output'] | null;
      readonly ticketType: CodecTypes['pg/text@1']['output'] | null;
      readonly hostId: CodecTypes['pg/int4@1']['output'] | null;
    };
    readonly User: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly emailVerified: CodecTypes['pg/timestamptz-temporal@1']['output'] | null;
      readonly image: CodecTypes['pg/text@1']['output'] | null;
      readonly username: CodecTypes['pg/text@1']['output'] | null;
      readonly name: CodecTypes['pg/text@1']['output'] | null;
      readonly googleId: CodecTypes['pg/text@1']['output'] | null;
      readonly accessToken: CodecTypes['pg/text@1']['output'] | null;
      readonly refreshToken: CodecTypes['pg/text@1']['output'] | null;
      readonly expiresAt: CodecTypes['pg/timestamptz-temporal@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly firstName: CodecTypes['pg/text@1']['output'] | null;
      readonly lastName: CodecTypes['pg/text@1']['output'] | null;
      readonly gsuiteId: CodecTypes['pg/text@1']['output'] | null;
      readonly branch: CodecTypes['pg/text@1']['output'] | null;
      readonly div: CodecTypes['pg/text@1']['output'] | null;
      readonly rollNo: CodecTypes['pg/text@1']['output'] | null;
      readonly admissionYear: CodecTypes['pg/int4@1']['output'] | null;
      readonly graduationYear: CodecTypes['pg/int4@1']['output'] | null;
      readonly tndUid: CodecTypes['pg/text@1']['output'] | null;
      readonly isExternal: CodecTypes['pg/bool@1']['output'];
      readonly isHost: CodecTypes['pg/bool@1']['output'];
      readonly memberOfClubId: CodecTypes['pg/int4@1']['output'] | null;
    };
    readonly Vault: {
      readonly uniqueId: CodecTypes['pg/int4@1']['output'];
      readonly eventId: CodecTypes['pg/int4@1']['output'];
      readonly userId: CodecTypes['pg/int4@1']['output'] | null;
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Club: {
      readonly clubId: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly branch: CodecTypes['pg/text@1']['input'] | null;
      readonly website: CodecTypes['pg/text@1']['input'] | null;
      readonly facebook: CodecTypes['pg/text@1']['input'] | null;
      readonly twitter: CodecTypes['pg/text@1']['input'] | null;
      readonly instagram: CodecTypes['pg/text@1']['input'] | null;
      readonly linkedin: CodecTypes['pg/text@1']['input'] | null;
      readonly facultyInchargeId: CodecTypes['pg/int4@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly Event: {
      readonly eventId: CodecTypes['pg/int4@1']['input'];
      readonly eventName: CodecTypes['pg/text@1']['input'];
      readonly eventDescription: CodecTypes['pg/text@1']['input'] | null;
      readonly aicteHr: CodecTypes['pg/int4@1']['input'] | null;
      readonly date: CodecTypes['pg/timestamptz-temporal@1']['input'] | null;
      readonly time: CodecTypes['pg/text@1']['input'] | null;
      readonly venue: CodecTypes['pg/text@1']['input'] | null;
      readonly hostId: CodecTypes['pg/int4@1']['input'] | null;
      readonly banner: CodecTypes['pg/text@1']['input'] | null;
      readonly brochure: CodecTypes['pg/text@1']['input'] | null;
    };
    readonly Ticket: {
      readonly ticketId: CodecTypes['pg/int4@1']['input'];
      readonly eventId: CodecTypes['pg/int4@1']['input'];
      readonly userId: CodecTypes['pg/int4@1']['input'] | null;
      readonly tnpUid: CodecTypes['pg/text@1']['input'] | null;
      readonly date: CodecTypes['pg/timestamptz-temporal@1']['input'] | null;
      readonly time: CodecTypes['pg/text@1']['input'] | null;
      readonly venue: CodecTypes['pg/text@1']['input'] | null;
      readonly eventName: CodecTypes['pg/text@1']['input'] | null;
      readonly ticketType: CodecTypes['pg/text@1']['input'] | null;
      readonly hostId: CodecTypes['pg/int4@1']['input'] | null;
    };
    readonly User: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly emailVerified: CodecTypes['pg/timestamptz-temporal@1']['input'] | null;
      readonly image: CodecTypes['pg/text@1']['input'] | null;
      readonly username: CodecTypes['pg/text@1']['input'] | null;
      readonly name: CodecTypes['pg/text@1']['input'] | null;
      readonly googleId: CodecTypes['pg/text@1']['input'] | null;
      readonly accessToken: CodecTypes['pg/text@1']['input'] | null;
      readonly refreshToken: CodecTypes['pg/text@1']['input'] | null;
      readonly expiresAt: CodecTypes['pg/timestamptz-temporal@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly firstName: CodecTypes['pg/text@1']['input'] | null;
      readonly lastName: CodecTypes['pg/text@1']['input'] | null;
      readonly gsuiteId: CodecTypes['pg/text@1']['input'] | null;
      readonly branch: CodecTypes['pg/text@1']['input'] | null;
      readonly div: CodecTypes['pg/text@1']['input'] | null;
      readonly rollNo: CodecTypes['pg/text@1']['input'] | null;
      readonly admissionYear: CodecTypes['pg/int4@1']['input'] | null;
      readonly graduationYear: CodecTypes['pg/int4@1']['input'] | null;
      readonly tndUid: CodecTypes['pg/text@1']['input'] | null;
      readonly isExternal: CodecTypes['pg/bool@1']['input'];
      readonly isHost: CodecTypes['pg/bool@1']['input'];
      readonly memberOfClubId: CodecTypes['pg/int4@1']['input'] | null;
    };
    readonly Vault: {
      readonly uniqueId: CodecTypes['pg/int4@1']['input'];
      readonly eventId: CodecTypes['pg/int4@1']['input'];
      readonly userId: CodecTypes['pg/int4@1']['input'] | null;
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;