// migrations/snapshots/829c6ebacc832dfa22f64b42b41aea7cbcf904286f48b99d1afaa06ce90d5158/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly User: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly emailVerified: CodecTypes['pg/timestamptz-temporal@1']['output'] | null;
      readonly image: CodecTypes['pg/text@1']['output'] | null;
      readonly username: CodecTypes['pg/text@1']['output'] | null;
      readonly name: CodecTypes['pg/text@1']['output'] | null;
      readonly googleId: CodecTypes['pg/text@1']['output'] | null;
      readonly accessToken: CodecTypes['pg/text@1']['output'] | null;
      readonly refreshToken: CodecTypes['pg/text@1']['output'] | null;
      readonly expiresAt: CodecTypes['pg/timestamptz-temporal@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly User: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly emailVerified: CodecTypes['pg/timestamptz-temporal@1']['input'] | null;
      readonly image: CodecTypes['pg/text@1']['input'] | null;
      readonly username: CodecTypes['pg/text@1']['input'] | null;
      readonly name: CodecTypes['pg/text@1']['input'] | null;
      readonly googleId: CodecTypes['pg/text@1']['input'] | null;
      readonly accessToken: CodecTypes['pg/text@1']['input'] | null;
      readonly refreshToken: CodecTypes['pg/text@1']['input'] | null;
      readonly expiresAt: CodecTypes['pg/timestamptz-temporal@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly user: {
      readonly accessToken: CodecTypes['pg/text@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly emailVerified: CodecTypes['pg/timestamptz-temporal@1']['output'] | null;
      readonly expiresAt: CodecTypes['pg/timestamptz-temporal@1']['output'] | null;
      readonly googleId: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly image: CodecTypes['pg/text@1']['output'] | null;
      readonly name: CodecTypes['pg/text@1']['output'] | null;
      readonly refreshToken: CodecTypes['pg/text@1']['output'] | null;
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly username: CodecTypes['pg/text@1']['output'] | null;
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly user: {
      readonly accessToken: CodecTypes['pg/text@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly emailVerified: CodecTypes['pg/timestamptz-temporal@1']['input'] | null;
      readonly expiresAt: CodecTypes['pg/timestamptz-temporal@1']['input'] | null;
      readonly googleId: CodecTypes['pg/text@1']['input'] | null;
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly image: CodecTypes['pg/text@1']['input'] | null;
      readonly name: CodecTypes['pg/text@1']['input'] | null;
      readonly refreshToken: CodecTypes['pg/text@1']['input'] | null;
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly username: CodecTypes['pg/text@1']['input'] | null;
    };
  };
};
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;